PT-2045 tidy up
